Fix exception handling when TNEF raises a ValueError
In an uncovered path exc.message was used, which is Python 2 only
syntax.

This has been updated.

As a suitable TNEF file is missing, the test makes heavy usage of mock.

If a suitable test file can be found, the test should be rewritten.

modified:   tests/test_cmdline.py
modified:   tnefparse/cmdline.py